Make it more setupable and extendable need to update docs

+++ setup/set_env.php
@@ -1,8 +1,19 @@
 <?php
 
+require_once __DIR__ . '/bootstrap.php';
+
+use Symfony\Component\Filesystem\Filesystem;
+
 try
 {
-    exec("cp env.example .env");
+    $filesystem = new Filesystem();
+    if($filesystem->exists('.env'))
+    {
+        echo sprintf("Env file exists already %s", __DIR__ . '/../.env');
+    } else {
+        echo sprintf("Copied .env file too root");
+        exec("cp env.example .env");
+    }
 }
 catch(\Exception $e)
 {
